Move code from machine-rdma.h to conv-rdma.C
