+++ src/conv-core/conv-rdma.C
@@ -12,19 +12,11 @@ static int zc_pup_handler_idx;
 
 // Methods required to keep the Nocopy Direct API functional on non-LRTS layers
 #if !CMK_USE_LRTS
-void CmiSetRdmaCommonInfo(void *info, const void *ptr, int size) {
-}
-
-int CmiGetRdmaCommonInfoSize() {
-  return 0;
-}
-
 void CmiSetNcpyAckSize(int ackSize) {}
 
 void CmiForwardNodeBcastMsg(int size, char *msg) {}
 
 void CmiForwardProcBcastMsg(int size, char *msg) {}
-
 #endif
 
 /****************************** Zerocopy Direct API For non-RDMA layers *****************************/
@@ -342,16 +334,8 @@ void zcPupGet(CmiNcpyBuffer &src, CmiNcpyBuffer &dest) {
   }
 }
 
-/* Support for Nocopy Direct API */
-void LrtsSetRdmaBufferInfo(void *info, const void *ptr, int size, unsigned short int mode);
-void LrtsIssueRget(NcpyOperationInfo *ncpyOpInfo);
-
-void LrtsIssueRput(NcpyOperationInfo *ncpyOpInfo);
-
-void LrtsDeregisterMem(const void *ptr, void *info, int pe, unsigned short int mode);
-
-#if CMK_REG_REQUIRED
-void LrtsInvokeRemoteDeregAckHandler(int pe, NcpyOperationInfo *ncpyOpInfo);
+#if CMK_USE_LRTS
+#include "machine-rdma.h"
 #endif
 
 
@@ -400,4 +384,78 @@ void CmiSetRdmaBufferInfo(void *info, const void *ptr, int size, unsigned short 
 /* Set the ack handler function used in the Direct API */
 void CmiSetDirectNcpyAckHandler(RdmaAckCallerFn fn){
   ncpyDirectAckHandlerFn = fn;
+}
+
+#if CMK_USE_CMA
+#include <unistd.h>
+#endif
+
+/* Support for Nocopy Direct API */
+typedef struct _cmi_common_rdma_info {
+#if CMK_USE_CMA
+  pid_t pid;
+#elif defined _MSC_VER
+  char empty;
+#endif
+} CmiCommonRdmaInfo_t;
+
+/* Set the generic converse/LRTS information */
+void CmiSetRdmaCommonInfo(void *info, const void *ptr, int size) {
+#if CMK_USE_CMA
+  CmiCommonRdmaInfo_t *cmmInfo = (CmiCommonRdmaInfo_t *)info;
+  cmmInfo->pid = getpid();
+#endif
+}
+
+int CmiGetRdmaCommonInfoSize() {
+#if CMK_USE_CMA
+  return sizeof(CmiCommonRdmaInfo_t);
+#else
+  return 0; // If CMK_USE_CMA is false, sizeof(CmiCommonRdmaInfo_t) is 1 (size of an empty structure in C++)
+            // However, 0 is returned since CMK_COMMON_NOCOPY_DIRECT_BYTES is set to 0 when CMK_USE_CMA is false
+            // because the offset (returned by CmiGetRdmaCommonInfoSize) should equal CMK_COMMON_NOCOPY_DIRECT_BYTES
+#endif
+}
+
+#if CMK_USE_CMA
+#include <unistd.h>
+#include <sys/uio.h> // for struct iovec
+extern int cma_works;
+int readShmCma(pid_t, char*, char*, size_t);
+int writeShmCma(pid_t, char *, char *, size_t);
+
+// These methods are also used by the generic layer implementation of the Direct API
+void CmiIssueRgetUsingCMA(
+  const void* srcAddr,
+  void *srcInfo,
+  int srcPe,
+  const void* destAddr,
+  void *destInfo,
+  int destPe,
+  size_t size) {
+
+  // get remote process id
+  CmiCommonRdmaInfo_t *remoteCommInfo = (CmiCommonRdmaInfo_t *)srcInfo;
+  pid_t pid = remoteCommInfo->pid;
+  readShmCma(pid, (char *)destAddr, (char *)srcAddr, size);
+}
+
+void CmiIssueRputUsingCMA(
+  const void* destAddr,
+  void *destInfo,
+  int destPe,
+  const void* srcAddr,
+  void *srcInfo,
+  int srcPe,
+  size_t size) {
+
+  // get remote process id
+  CmiCommonRdmaInfo_t *remoteCommInfo = (CmiCommonRdmaInfo_t *)destInfo;
+  pid_t pid = remoteCommInfo->pid;
+  writeShmCma(pid, (char *)srcAddr, (char *)destAddr, size);
+}
+#endif
+
+void CmiInvokeNcpyAck(void *ack) {
+  ncpyDirectAckHandlerFn(ack);
 }
